requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core==1.26.3
azure-functions==1.13.2
azure-storage-blob==12.15.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
idna==3.4
isodate==0.6.1
pycparser==2.21
requests==2.28.2
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.14